[KFI] Refactor: rename methods and members. Prefer "predicate" instead of "query".
