+++ src/ContentRepository/ContentRepository.csproj
@@ -168,7 +168,6 @@
     <Compile Include="Configuration\Services.cs" />
     <Compile Include="Configuration\Versioning.cs" />
     <Compile Include="Linq\CaseInsensitiveFilterVisitor.cs" />
-    <Compile Include="Linq\CQVisitor_NEW.cs" />
     <Compile Include="Packaging\Dependency.cs" />
     <Compile Include="Packaging\ExecutionContext.cs" />
     <Compile Include="Packaging\InstallResult.cs" />
@@ -266,7 +265,7 @@
     <Compile Include="Json\JsonDynamicFieldHelper.cs" />
     <Compile Include="Linq\ChildrenContentSet.cs" />
     <Compile Include="Linq\ContentSet.cs" />
-    <Compile Include="Linq\CQVisitor.cs" />
+    <Compile Include="Linq\SnLinqVisitor.cs" />
     <Compile Include="Linq\ExecutorVisitor.cs" />
     <Compile Include="Linq\FilteredEnumerable.cs" />
     <Compile Include="Linq\LinqChildrenEnumerator.cs" />
